Fix tango admin failure issue

Closes #1128

See merge request tango-controls/cppTango!1109

+++ tests/environment/kill_server.sh.cmake
@@ -7,5 +7,10 @@ set -o nounset
 
 pids=$(tr '\n' ' ' <"${TANGO_TEST_CASE_DIRECTORY}/server_pids")
 
+if [[ -z "$pids" ]]; then
+    echo "No device servers to kill"
+    exit
+fi
+
 echo "Killing PIDs: $pids"
 kill -s TERM --timeout 5000 KILL -- $pids &>/dev/null
